Rename scil_compute_mean_fixel_adf*
