+++ scripts/legacy/scil_compute_bundle_mean_std_per_point.py
@@ -2,7 +2,7 @@
 # -*- coding: utf-8 -*-
 
 from scilpy.io.deprecator import deprecate_script
-from scripts.scil_tractogram_print_info import main as new_main
+from scripts.scil_bundle_mean_std import main as new_main
 
 
 DEPRECATION_MSG = """
