+++ scripts/legacy/scil_compute_mean_fixel_afd_from_bundles.py
@@ -2,16 +2,17 @@
 # -*- coding: utf-8 -*-
 
 from scilpy.io.deprecator import deprecate_script
-from scripts.scil_bundle_mean_std import main as new_main
+from scripts.scil_bundle_mean_fixel_afd import main as new_main
 
 
 DEPRECATION_MSG = """
-This script has been renamed scil_bundle_mean_std.py.
+This script has been renamed scil_bundle_mean_fixel_afd.py.
 Please change your existing pipelines accordingly.
 """
 
 
-@deprecate_script("scil_compute_bundle_mean_std.py", DEPRECATION_MSG, '1.7.0')
+@deprecate_script("scil_compute_fixel_afd_from_bundles.py",
+                  DEPRECATION_MSG, '1.7.0')
 def main():
     new_main()
