+++ scripts/tests/test_bundle_mean_fixel_afd.py
@@ -12,19 +12,16 @@ tmp_dir = tempfile.TemporaryDirectory()
 
 
 def test_help_option(script_runner):
-    ret = script_runner.run('scil_compute_mean_fixel_afd_from_bundles.py',
-                            '--help')
+    ret = script_runner.run('scil_bundle_fixel_afd.py', '--help')
     assert ret.success
 
 
 def test_execution_processing(script_runner):
     os.chdir(os.path.expanduser(tmp_dir.name))
-    in_tracking = os.path.join(get_home(), 'processing',
-                               'tracking.trk')
+    in_tracking = os.path.join(get_home(), 'processing', 'tracking.trk')
     in_fodf = os.path.join(get_home(), 'processing',
                            'fodf_descoteaux07.nii.gz')
-    ret = script_runner.run('scil_compute_mean_fixel_afd_from_bundles.py',
-                            in_tracking, in_fodf,
+    ret = script_runner.run('scil_bundle_fixel_afd.py', in_tracking, in_fodf,
                             'afd_test.nii.gz',
                             '--sh_basis', 'descoteaux07', '--length_weighting')
     assert ret.success
